Move field parsing into the column + simplify the crate structure
