--- readyset-vitess-data/src/vitess/mod.rs
@@ -1,7 +1,0 @@
-mod column;
-mod schema_cache;
-mod table;
-
-pub use column::Column;
-pub use schema_cache::SchemaCache;
-pub use table::Table;
