+++ readyset-vitess-data/tests/field_parsing.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use bit_vec::BitVec;
 use readyset_data::DfValue;
-use readyset_vitess_data::vstream_value_to_noria_value;
+use readyset_vitess_data::Column;
 use vitess_grpc::query::Type;
 
 #[test]
@@ -16,107 +16,118 @@ fn int_parsing() {
     ]
     .iter()
     .for_each(|t| {
-        let res = vstream_value_to_noria_value("42".as_bytes(), *t, None);
+        let column = Column::new("test", *t, "int", 0);
+        let res = column.vstream_value_to_noria_value("42".as_bytes());
         assert!(res.is_ok());
         assert_eq!(res.unwrap(), DfValue::Int(42));
 
-        let res = vstream_value_to_noria_value("hello".as_bytes(), *t, None);
+        let res = column.vstream_value_to_noria_value("hello".as_bytes());
         assert!(res.is_err());
     });
 }
 
 #[test]
 fn date_parsing() {
-    let res = vstream_value_to_noria_value("2023-07-24".as_bytes(), Type::Date, None);
+    let column = Column::new("test", Type::Date, "date", 0);
+    let res = column.vstream_value_to_noria_value("2023-07-24".as_bytes());
     assert!(res.is_ok());
     assert_eq!(res.unwrap(), "2023-07-24".try_into().unwrap());
 
-    let res = vstream_value_to_noria_value("hello".as_bytes(), Type::Date, None);
+    let res = column.vstream_value_to_noria_value("hello".as_bytes());
     assert!(res.is_err());
 }
 
 #[test]
 fn datetime_parsing() {
-    let res = vstream_value_to_noria_value("2023-07-24 12:34:56".as_bytes(), Type::Datetime, None);
+    let column = Column::new("test", Type::Datetime, "datetime", 0);
+    let res = column.vstream_value_to_noria_value("2023-07-24 12:34:56".as_bytes());
     assert!(res.is_ok());
     assert_eq!(res.unwrap(), "2023-07-24 12:34:56".try_into().unwrap());
 
-    let res = vstream_value_to_noria_value("hello".as_bytes(), Type::Datetime, None);
+    let res = column.vstream_value_to_noria_value("hello".as_bytes());
     assert!(res.is_err());
 }
 
 #[test]
 fn time_parsing() {
-    let res = vstream_value_to_noria_value("-12:34:56".as_bytes(), Type::Time, None);
+    let column = Column::new("test", Type::Time, "time", 0);
+    let res = column.vstream_value_to_noria_value("-12:34:56".as_bytes());
     assert!(res.is_ok());
     assert_eq!(res.unwrap(), "-12:34:56".try_into().unwrap());
 
-    let res = vstream_value_to_noria_value("12:34:56".as_bytes(), Type::Time, None);
+    let res = column.vstream_value_to_noria_value("12:34:56".as_bytes());
     assert!(res.is_ok());
     assert_eq!(res.unwrap(), "12:34:56".try_into().unwrap());
 
-    let res = vstream_value_to_noria_value("120:34:56".as_bytes(), Type::Time, None);
+    let res = column.vstream_value_to_noria_value("120:34:56".as_bytes());
     assert!(res.is_ok());
     assert_eq!(res.unwrap(), "120:34:56".try_into().unwrap());
 
-    let res = vstream_value_to_noria_value("hello".as_bytes(), Type::Time, None);
+    let res = column.vstream_value_to_noria_value("hello".as_bytes());
     assert!(res.is_err());
 }
 
 #[test]
 fn timestamp_parsing() {
-    let res = vstream_value_to_noria_value(
-        "2038-01-19 03:14:07.499999".as_bytes(),
-        Type::Timestamp,
-        None,
-    );
+    let column = Column::new("test", Type::Timestamp, "timestamp", 0);
+    let res = column.vstream_value_to_noria_value("2038-01-19 03:14:07.499999".as_bytes());
     assert!(res.is_ok());
     assert_eq!(
         res.unwrap(),
         "2038-01-19 03:14:07.499999".try_into().unwrap()
     );
 
-    let res = vstream_value_to_noria_value("hello".as_bytes(), Type::Timestamp, None);
+    let res = column.vstream_value_to_noria_value("hello".as_bytes());
     assert!(res.is_err());
 }
 
 #[test]
 fn year_parsing() {
-    let res = vstream_value_to_noria_value("2023".as_bytes(), Type::Year, None);
+    let column = Column::new("test", Type::Year, "year", 0);
+    let res = column.vstream_value_to_noria_value("2023".as_bytes());
     assert!(res.is_ok());
     assert_eq!(res.unwrap(), DfValue::Int(2023));
 
-    let res = vstream_value_to_noria_value("hello".as_bytes(), Type::Year, None);
+    let res = column.vstream_value_to_noria_value("hello".as_bytes());
     assert!(res.is_err());
 }
 
 #[test]
 fn text_parsing() {
-    let res = vstream_value_to_noria_value("Hello, world!".as_bytes(), Type::Text, None);
+    let column = Column::new("test", Type::Text, "text", 0);
+    let res = column.vstream_value_to_noria_value("Hello, world!".as_bytes());
     assert!(res.is_ok());
     assert_eq!(res.unwrap(), DfValue::Text("Hello, world!".into()));
 
-    let res = vstream_value_to_noria_value("Lorem ipsum.".as_bytes(), Type::Char, None);
+    let column = Column::new("test", Type::Char, "char(100)", 0);
+    let res = column.vstream_value_to_noria_value("Lorem ipsum.".as_bytes());
     assert!(res.is_ok());
     assert_eq!(res.unwrap(), DfValue::Text("Lorem ipsum.".into()));
 
-    let res = vstream_value_to_noria_value("Short text".as_bytes(), Type::Varchar, None);
+    let column = Column::new("test", Type::Varchar, "varchar(100)", 0);
+    let res = column.vstream_value_to_noria_value("Short text".as_bytes());
     assert!(res.is_ok());
     assert_eq!(res.unwrap(), DfValue::Text("Short text".into()));
 
     let json = "{\"x\": 123}";
-    let res = vstream_value_to_noria_value(json.as_bytes(), Type::Json, None);
+    let column = Column::new("test", Type::Json, "json", 0);
+    let res = column.vstream_value_to_noria_value(json.as_bytes());
     assert!(res.is_ok());
     assert_eq!(res.unwrap(), DfValue::Text(json.into()));
 }
 
 #[test]
 fn binary_parsing() {
+    let binary_types = [
+        (Type::Blob, "blob"),
+        (Type::Binary, "binary"),
+        (Type::Varbinary, "varbinary"),
+    ];
     let raw_value = "binary data".as_bytes();
-    let binary_types = [Type::Blob, Type::Binary, Type::Varbinary];
 
-    for binary_type in binary_types.iter() {
-        let res = vstream_value_to_noria_value(raw_value, *binary_type, None);
+    for (binary_type, type_name) in binary_types.iter() {
+        let column = Column::new("test", *binary_type, type_name, 0);
+        let res = column.vstream_value_to_noria_value(raw_value);
         assert!(res.is_ok());
         assert_eq!(
             res.unwrap(),
@@ -127,7 +138,8 @@ fn binary_parsing() {
 
 #[test]
 fn bit_parsing_single_byte() {
-    let res = vstream_value_to_noria_value(&[0b10101010], Type::Bit, None);
+    let column = Column::new("test", Type::Bit, "bit(8)", 0);
+    let res = column.vstream_value_to_noria_value(&[0b10101010]);
     assert!(res.is_ok());
     let expected = BitVec::from_bytes(&[0b10101010]);
     assert_eq!(res.unwrap(), DfValue::BitVector(Arc::new(expected)));
@@ -135,7 +147,8 @@ fn bit_parsing_single_byte() {
 
 #[test]
 fn bit_parsing_multiple_bytes() {
-    let res = vstream_value_to_noria_value(&[0b10101010, 0b11001100], Type::Bit, None);
+    let column = Column::new("test", Type::Bit, "bit(8)", 0);
+    let res = column.vstream_value_to_noria_value(&[0b10101010, 0b11001100]);
     assert!(res.is_ok());
     let expected = BitVec::from_bytes(&[0b10101010, 0b11001100]);
     assert_eq!(res.unwrap(), DfValue::BitVector(Arc::new(expected)));
@@ -143,7 +156,8 @@ fn bit_parsing_multiple_bytes() {
 
 #[test]
 fn empty_field() {
-    let res = vstream_value_to_noria_value(&[], Type::Int8, None);
+    let column = Column::new("test", Type::Bit, "bit(8)", 0);
+    let res = column.vstream_value_to_noria_value(&[]);
     assert!(res.is_ok());
     assert_eq!(res.unwrap(), DfValue::None);
 }
